move validatePlugins to PluginUtil
